backend/fastapi/app/models/schemas.py
from pydantic import BaseModel


class HealthResponse(BaseModel):
    status: str


# Auth schemas
class UserCreate(BaseModel):
    username: str
    password: str


class UserLogin(BaseModel):
    username: str
    password: str


class Token(BaseModel):
    access_token: str
    token_type: str


class TokenData(BaseModel):
    username: str | None = None


class UserResponse(BaseModel):
    id: int
    username: str
    created_at: str
