backend/fastapi/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.5.0
pydantic-settings>=2.0.0
